packages/kittyai/requirements.txt
openai
kitty
kittens
python-dotenv
